just removed object c app delegate

+++ local/Common/Cluster/QCluster+Annotation.m
@@ -14,7 +14,7 @@
 
 -(NSString*)subtitle
 {
-  return [NSString stringWithFormat:@"%d objects here", self.objectsCount];
+  return [NSString stringWithFormat:@"%ld objects here", (long)self.objectsCount];
 }
 
 @end
